DOCUMENT_UNDERSTANDING/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
requests==2.28.2
Pillow==9.5.0
google-generativeai==0.7.2
PyMuPDF==1.20.2
